split_channels: moved from `utils` to `channels`
- moved method to more sensible module
- a bunch of small fixes for GEO channel parsing
