+++ gwsumm/plot/range.py
@@ -34,7 +34,7 @@ from .. import globalv
 from .registry import (get_plot, register_plot)
 from ..data import (get_range_channel, get_range, get_timeseries)
 from ..segments import get_segments
-from ..utils import split_channels
+from ..channels import split as split_channels
 
 __author__ = 'Duncan Macleod <duncan.macleod@ligo.org>'
